refactor: Move github tools to separate file
